Pull resources handling into its own file.
Conflicts:
	internal/v5/api.go
